Make the top level module graph easier to access - Now has a main - Needs cleanup on file naming
